+++ TabLayout/TabLayout.js
@@ -19,7 +19,7 @@ import {getTargetByDirectionFromElement} from '@enact/spotlight/src/target';
 import SpotlightContainerDecorator from '@enact/spotlight/SpotlightContainerDecorator';
 import {Changeable} from '@enact/ui/Changeable';
 import {Cell, Layout} from '@enact/ui/Layout';
-import {scaleToRem} from '@enact/ui/resolution';
+import ri, {scaleToRem} from '@enact/ui/resolution';
 import Toggleable from '@enact/ui/Toggleable';
 import Touchable from '@enact/ui/Touchable';
 import ViewManager from '@enact/ui/ViewManager';
@@ -36,11 +36,42 @@ import Tab from './Tab';
 import componentCss from './TabLayout.module.less';
 import popupTabLayoutComponentCss from '../PopupTabLayout/PopupTabLayout.module.less';
 
+const MAX_TABS_BEFORE_VERTICAL_SCROLLING = 8;
+const TAB_SPACING = 48;
+
 const TabLayoutContext = createContext(null);
 
 const TouchableCell = Touchable(Cell);
 
 const isTouchMode = () => (getLastInputType() === 'touch');
+
+const getHorizontalTabWidth = (dataSize, size, tabSize) => {
+	let widths;
+
+	if (tabSize) {
+		return tabSize;
+	}
+
+	if (size === 'small') {
+		widths = [540, 420, 420];
+	} else {
+		widths = [852, 672, 552];
+	}
+
+	if (dataSize < 5) {
+		return widths[0];
+	} else if (dataSize < 6) {
+		return widths[1];
+	} else {
+		return widths[2];
+	}
+};
+
+const isHorizontalScrollableTabs = (dataSize, size, tabSize) => {
+	const totalTabsWidth = dataSize * getHorizontalTabWidth(dataSize, size, tabSize) + TAB_SPACING * (dataSize - 1);
+
+	return (typeof window !== 'undefined' && window?.screen?.width) ? window.screen.width < ri.scale(totalTabsWidth) : false;
+};
 
 /**
  * Tabbed Layout component.
@@ -79,6 +110,16 @@ const TabLayoutBase = kind({
 		 * @private
 		 */
 		anchorTo: PropTypes.oneOf(['left', 'right', 'start', 'end']),
+
+		/**
+		 * Prevents the tab list from automatically collapsing when the screen orientation changes to portrait mode.
+		 *
+		 * Only applies to `orientation="vertical".
+		 *
+		 * @type {Boolean}
+		 * @public
+		 */
+		blockCollapseOnPortrait: PropTypes.bool,
 
 		/**
 		 * Collection of {@link limestone/TabLayout.Tab|Tabs} to render.
@@ -152,6 +193,18 @@ const TabLayoutBase = kind({
 		index: PropTypes.number,
 
 		/**
+		 * The offset of the tabs area from the left and right edges of the screen.
+		 * This option is only applicable when `orientation` is set to `horizontal`.
+		 * The default value set to `36` to have the right position of the tabs in the normal panel.
+		 * If the tabs are supposed to scroll, this value should be set to `132` and the panel should have no padding.
+		 *
+		 * @type {Number}
+		 * @default 36
+		 * @public
+		 */
+		offset: PropTypes.number,
+
+		/**
 		 * Called when the tabs are collapsed.
 		 *
 		 * @type {Function}
@@ -197,12 +250,33 @@ const TabLayoutBase = kind({
 		orientation: PropTypes.oneOf(['horizontal', 'vertical']),
 
 		/**
+		 * The index of the primary tab.
+		 * When this prop is set, the initial focus will be on the primary tab when rendered.
+		 * Also, when pressing the back key from other tabs, the focus will be moved to the primary tab.
+		 * If the `primaryIndex` is not provided, the initial focus will default to the first tab during the initial render.
+		 *
+		 * @type {Number}
+		 * @default null
+		 * @public
+		 */
+		primaryIndex: PropTypes.number,
+
+		/**
 		 * Indicates the content's text direction is right-to-left.
 		 *
 		 * @type {Boolean}
 		 * @private
 		 */
 		rtl: PropTypes.bool,
+
+		/**
+		 * The size of the horizontal tab.
+		 *
+		 * @type {('small'|'large')}
+		 * @default 'large'
+		 * @public
+		 */
+		size: PropTypes.oneOf(['small', 'large']),
 
 		/**
 		 * Assign a custom size to horizontal tabs.
@@ -235,15 +309,18 @@ const TabLayoutBase = kind({
 		dimensions: {
 			tabs: {
 				collapsed: 216,
-				normal: 882
+				normal: 888
 			},
 			content: {
 				expanded: null,
 				normal: null
 			}
 		},
-		index: 0,
+		index: null,
+		primaryIndex: null,
+		offset: 36,
 		orientation: 'vertical',
+		size: 'large',
 		type: 'normal'
 	},
 
@@ -279,19 +356,23 @@ const TabLayoutBase = kind({
 		},
 		onKeyUp: (ev, props) => {
 			const {keyCode, target} = ev;
-			const {anchorTo, collapsed, orientation, 'data-spotlight-id': spotlightId, rtl, type} = props;
+			const {anchorTo, collapsed, orientation, primaryIndex, 'data-spotlight-id': spotlightId, rtl, type} = props;
 			const popupPanelRef = document.querySelector(`[data-spotlight-id='${spotlightId}'] .${popupTabLayoutComponentCss.panel}`);
 			const tabLayoutContentRef = document.querySelector(`[data-spotlight-id='${spotlightId}'] .${componentCss.content}`);
+			const tabsExpandedSpotlightId = `${spotlightId}-tabs-expanded`;
 
 			if (forwardWithPrevent('onKeyUp', ev, props) && is('cancel')(keyCode)) {
 				if ((type === 'popup' && popupPanelRef?.contains(target) && popupPanelRef?.dataset.index === '0') || (type === 'normal' && !Spotlight.getPointerMode() && tabLayoutContentRef?.contains(target))) {
 					if (collapsed) {
 						forward('onExpand', ev, props);
 					}
-					Spotlight.focus(`[data-spotlight-id='${spotlightId}-tabs-expanded']`);
+					Spotlight.focus(`[data-spotlight-id='${tabsExpandedSpotlightId}']`);
+					ev.stopPropagation();
+				} else if (primaryIndex !== null) {
+					Spotlight.focus(`[data-spotlight-id='${tabsExpandedSpotlightId}-primary-tab']`);
 					ev.stopPropagation();
 				}
-			} else if (is('enter')(keyCode) && !collapsed && document.querySelector(`[data-spotlight-id='${spotlightId}-tabs-expanded']`).contains(target) && target.tagName !== 'INPUT') {
+			} else if (is('enter')(keyCode) && !collapsed && document.querySelector(`[data-spotlight-id='${tabsExpandedSpotlightId}']`).contains(target) && target.tagName !== 'INPUT') {
 				Spotlight.setPointerMode(false);
 
 				let moveTo;
@@ -365,10 +446,33 @@ const TabLayoutBase = kind({
 			`anchor${cap(anchorTo)}`,
 			orientation
 		),
-		style: ({dimensions, orientation, style}) => ({
-			...style,
-			'--tablayout-expand-collapse-diff': ((orientation === 'vertical') ? scaleToRem(dimensions.tabs.normal - dimensions.tabs.collapsed) : 0)
-		}),
+		scrollable: ({children, orientation, size, tabSize}) => {
+			const isVertical = orientation === 'vertical';
+			return isVertical ?
+				(children.length > MAX_TABS_BEFORE_VERTICAL_SCROLLING) :
+				isHorizontalScrollableTabs(children.length, size, tabSize);
+		},
+		style: ({children, dimensions, offset, orientation, size, style, tabSize}) => {
+			const isVertical = orientation === 'vertical';
+			const scrollable = isVertical ?
+				(children.length > MAX_TABS_BEFORE_VERTICAL_SCROLLING) :
+				isHorizontalScrollableTabs(children.length, size, tabSize);
+			const tabSizeValue = !isVertical ? getHorizontalTabWidth(children.length, size, tabSize) : null;
+			const totalTabsWidth = ri.scaleToRem(tabSizeValue * children.length + TAB_SPACING * (children.length - 1));
+
+			return {
+				...style,
+				'--offset': scaleToRem(offset),
+				'--scrollable': scrollable ? '1' : '0',
+				'--tabs-width': !isVertical ? totalTabsWidth : '100%',
+				'--tab-width': scaleToRem(tabSizeValue),
+				'--tablayout-expand-collapse-diff': ((orientation === 'vertical') ? scaleToRem(dimensions.tabs.normal - dimensions.tabs.collapsed) : 0)
+			};
+		},
+		index: ({index, primaryIndex}) => {
+			// If `index` is not provided, it defaults to `primaryIndex` or 0.
+			return index ?? primaryIndex ?? 0;
+		},
 		tabOrientation: ({orientation}) => orientation === 'vertical' ? 'horizontal' : 'vertical',
 		tabs: ({children}) => {
 			const tabs = mapAndFilterChildren(children, (child) => (
@@ -380,11 +484,14 @@ const TabLayoutBase = kind({
 		}
 	},
 
-	render: ({children, collapsed, css, 'data-spotlight-id': spotlightId, dimensions, handleClick, handleEnter, handleFlick, handleFocus, handleTabsTransitionEnd, index, onCollapse, onSelect, orientation, tabOrientation, tabSize, tabs, type, ...rest}) => {
+	render: ({children, collapsed, css, 'data-spotlight-id': spotlightId, primaryIndex, dimensions, handleClick, handleEnter, handleFlick, handleFocus, handleTabsTransitionEnd, index, onCollapse, onSelect, orientation, scrollable, size, tabOrientation, tabs, type, ...rest}) => {
 		delete rest.anchorTo;
+		delete rest.blockCollapseOnPortrait;
 		delete rest.onExpand;
+		delete rest.offset;
 		delete rest.onTabAnimationEnd;
 		delete rest.rtl;
+		delete rest.tabSize;
 
 		const contentSize = (collapsed ? dimensions.content.expanded : dimensions.content.normal);
 		const isVertical = orientation === 'vertical';
@@ -394,6 +501,7 @@ const TabLayoutBase = kind({
 		// Props that are shared between both of the rendered TabGroup components
 		const tabGroupProps = {
 			css,
+			primaryIndex,
 			onClick: (collapsed ? handleClick : null),
 			onFocus: (collapsed ? handleFocus : null),
 			onFocusTab: onSelect,
@@ -405,14 +513,15 @@ const TabLayoutBase = kind({
 
 		// In vertical orientation, render two sets of tabs, one just icons, one with icons and text.
 		return (
-			<TabLayoutContext.Provider value={handleEnter}>
+			<TabLayoutContext value={handleEnter}>
 				<Layout {...rest} orientation={tabOrientation} data-spotlight-id={spotlightId}>
 					<Cell className={css.tabs} shrink onTransitionEnd={handleTabsTransitionEnd}>
 						<TabGroup
 							{...tabGroupProps}
 							collapsed={isVertical}
+							scrollable={scrollable}
+							size={size}
 							spotlightId={getTabsSpotlightId(spotlightId, isVertical)}
-							tabSize={!isVertical ? tabSize : null}
 							spotlightDisabled={!collapsed && isVertical}
 						/>
 					</Cell>
@@ -422,6 +531,7 @@ const TabLayoutBase = kind({
 					>
 						<TabGroup
 							{...tabGroupProps}
+							scrollable={scrollable}
 							spotlightId={getTabsSpotlightId(spotlightId, false)}
 							spotlightDisabled={collapsed}
 						/>
@@ -439,7 +549,7 @@ const TabLayoutBase = kind({
 						{children}
 					</ContentCell>
 				</Layout>
-			</TabLayoutContext.Provider>
+			</TabLayoutContext>
 		);
 	}
 });