+++ tests/ui/apps/TabLayout/HorizontalTabs/TabLayoutHorizontalTabs-View.js
@@ -10,15 +10,15 @@ import {tabsForScroll} from '../TabComponents';
 spotlight.setPointerMode(false);
 
 const app = (props) => (
-	<div {...props} style={{display: 'flex', flexDirection: 'row', overflow: 'scroll'}}>
-		<TabLayout
-			id="tabLayout"
-			orientation="horizontal"
-			tabSize={900}
-		>
-			{tabsForScroll}
-		</TabLayout>
-	</div>
+	<TabLayout
+		{...props}
+		id="tabLayout"
+		offset={0}
+		orientation="horizontal"
+		tabSize={900}
+	>
+		{tabsForScroll}
+	</TabLayout>
 );
 
 export default ThemeDecorator(app);