+++ tests/ui/specs/TabLayout/HorizontalTabs/TabLayoutHorizontalTabs-specs.js
@@ -16,6 +16,7 @@ describe('TabLayout', function () {
 						await Page.spotlightRight();
 					}
 					// Step 2: Wait for the last view
+					await Page.delay(1000);
 					await (await Page.tabLayout.view(15)).waitForExist();
 					// Step 3: Check if the last tab is selected
 					expect(await Page.tabLayout.isSelectedTab(14)).toBe(true);
@@ -30,6 +31,7 @@ describe('TabLayout', function () {
 						await Page.spotlightRight();
 					}
 					// Step 2: Wait for the last view
+					await Page.delay(1000);
 					await (await Page.tabLayout.view(15)).waitForExist();
 					// Step 3: Check if the last tab is selected
 					expect(await Page.tabLayout.isSelectedTab(14)).toBe(true);
@@ -40,6 +42,7 @@ describe('TabLayout', function () {
 						await Page.spotlightLeft();
 					}
 					// Step 6: Wait for the first view
+					await Page.delay(1000);
 					await (await Page.tabLayout.view(1)).waitForExist();
 					// Step 7: Check if the first tab is selected
 					expect(await Page.tabLayout.isSelectedTab(0)).toBe(true);
@@ -51,33 +54,27 @@ describe('TabLayout', function () {
 			describe('pointer interaction', function () {
 				it('should scroll to the last Spottable tab', async function () {
 					const tabs = await (await Page.tabLayout.tabsFromScroller());
-					// Step 1: Scroll to the 5th tab to make it visible
-					tabs[4].scrollIntoView();
-					// Step 2: Move the pointer to the 5th tab
-					tabs[4].moveTo();
-					// Step 3: Wait for the scrolling animation to complete
+					// Step 1: Move the pointer to the 6th tab
+					await tabs[5].moveTo({xOffset: 200, yOffset: 0});
+					// Step 2: Wait for the scrolling animation to complete
 					await Page.delay(4000);
-					// Step 4: Verify that the 15th tab is visible within the viewport
+					// Step 3: Verify that the 15th tab is visible within the viewport
 					expect(await tabs[14].isDisplayed({withinViewport: true})).toBe(true);
 				});
 
 				it('should scroll to the first Spottable tab', async function () {
 					const tabs = await (await Page.tabLayout.tabsFromScroller());
-					// Step 1: Scroll to the 5th tab to make it visible
-					tabs[4].scrollIntoView();
-					// Step 2: Move the pointer to the 5th tab
-					tabs[4].moveTo();
-					// Step 3: Wait for the scrolling animation to complete
+					// Step 1: Move the pointer to the 6th tab
+					await tabs[5].moveTo({xOffset: 200, yOffset: 0});
+					// Step 2: Wait for the scrolling animation to complete
 					await Page.delay(4000);
-					// Step 4: Verify that the 15th tab is visible within the viewport
+					// Step 3: Verify that the 15th tab is visible within the viewport
 					expect(await tabs[14].isDisplayed({withinViewport: true})).toBe(true);
-					// Step 5: Scroll to the 11th tab to make it visible
-					tabs[10].scrollIntoView();
-					// Step 6: Move the pointer to the 11th tab
-					tabs[10].moveTo();
-					// Step 7: Wait for the scrolling animation to complete
+					// Step 4: Move the pointer to the 10th tab
+					await tabs[9].moveTo({xOffset: -200, yOffset: 0});
+					// Step 5: Wait for the scrolling animation to complete
 					await Page.delay(4000);
-					// Step 8: Verify that the 1st tab is visible within the viewport
+					// Step 6: Verify that the 1st tab is visible within the viewport
 					expect(await tabs[0].isDisplayed({withinViewport: true})).toBe(true);
 				});
 			});