+++ tests/ui/specs/VirtualList/VirtualList/ElementSize/VirtualList-QWTC-2320-specs.js
@@ -5,13 +5,13 @@ describe('Scrollbar Size', function () {
 		await Page.open();
 	});
 
-	it('should have 30px padding for the top and bottom of the scrollbar[QWTC-2320], [QWTC-2051]', async function () {
+	it('should have 12px padding for the top and bottom of the scrollbar[QWTC-2320], [QWTC-2051]', async function () {
 		// Verify: The scrollbar size fit to the size of the list.
 		expect((await Page.getListRect()).height).toBe((await Page.getVerticalScrollbarRect()).height);
-		// There is 30px padding of the top and bottom of the Scrollbar.
+		// There is 12px padding of the top and bottom of the Scrollbar.
 		// Check top Padding
-		expect((await Page.getVerticalScrollbarTrackRect()).top - (await Page.getVerticalScrollbarRect()).top).toBe(15);
+		expect((await Page.getVerticalScrollbarTrackRect()).top - (await Page.getVerticalScrollbarRect()).top).toBe(6);
 		// Check bottom Padding
-		expect((await Page.getVerticalScrollbarRect()).bottom - (await Page.getVerticalScrollbarTrackRect()).bottom).toBe(15);
+		expect((await Page.getVerticalScrollbarRect()).bottom - (await Page.getVerticalScrollbarTrackRect()).bottom).toBe(6);
 	});
 });